requirements.txt
streamlit==1.25.0
beautifulsoup4==4.12.3
newspaper3k==0.2.8
langchain==0.0.200
groq==0.9.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
chromadb==0.4.3chrom
